Migrate combobox from /docs/components/ to /components/
Move all examples (Basic, Form, Grouped) to the new reference page,
delete the old documentation page, and update all routes, sidebar
navigation, home page links, and E2E tests.
